add A

+++ AGC015/A.cpp
@@ -14,8 +14,8 @@ using namespace std;
 typedef long long ll;
 
 int main() {
-    int n, a, b; cin >> n >> a >> b;
-    int ans = 0;
+    ll n, a, b; cin >> n >> a >> b;
+    ll ans = 0;
     if(n == 1) {
         if(a != b) ans = 0;
         else ans = 1;
